Cleaned up compiler warning in STKDiscretization.
git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@944 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 65248fa255ec6401c079e8f785b1b3d5070237ef

+++ src/Peridigm_STKDiscretization.cpp
@@ -99,7 +99,7 @@ PeridigmNS::STKDiscretization::STKDiscretization(const Teuchos::RCP<const Epetra
   int* neighborhood = decomp.neighborhood.get();
   int neighborhoodIndex = 0;
   int numPointsWithZeroNeighbors = 0;
-  for(int i=0 ; i<decomp.numPoints ; ++i){
+  for(size_t i=0 ; i<decomp.numPoints ; ++i){
     int numNeighbors = neighborhood[neighborhoodIndex];
     if(numNeighbors > 0){
       numMyElements++;
